Fix pyinstaller import error, rename pairinteraction_gui/pairinteraction to pairinteraction_gui/pairinteraction_app
