+++ .flake8
@@ -2,4 +2,4 @@
 max-line-length = 120
 extend-ignore = E203, E225, E226, E266, E741
 per-file-ignores =
-    pairinteraction_gui/pairinteraction/app.py:E402
+    pairinteraction_gui/pairinteraction_app/app.py:E402
