+++ apple/launcher.sh
@@ -2,5 +2,5 @@
 
 # http://stackoverflow.com/questions/26465240/how-to-execute-a-shell-script-from-within-an-app-and-show-it-in-the-terminal
 
-scriptPath=$(dirname "$0")'/../Resources/pairinteraction_gui/pairinteraction'
+scriptPath=$(dirname "$0")'/../Resources/pairinteraction_gui/pairinteraction_gui'
 osascript -e 'tell app "Terminal" to do script "\"'"$scriptPath"'\"; exit"'
