+++ pairinteraction_gui/pipy/__init__.py
@@ -9,7 +9,5 @@ __all__ = [
     "atom_from_config",
 ]
 
-from pairinteraction import pireal, picomplex  # noqa, imported here, used in the submodules
-
 from pipy.config import Config
 from pipy.atom import Atom, AtomOne, AtomTwo, atom_from_config
