+++ setup.py.in
@@ -25,7 +25,7 @@ setup(
         "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
     ],
     packages = find_namespace_packages(include=[
-        "pairinteraction", "pairinteraction_gui", "pairinteraction_gui.pairinteraction", "pairinteraction_gui.pipy", "pairinteraction_gui.pairinteraction.*"
+        "pairinteraction", "pairinteraction_gui", "pairinteraction_gui.pairinteraction_app", "pairinteraction_gui.pipy", "pairinteraction_gui.pairinteraction_app.*"
     ]),
     package_data = {
         "pairinteraction" : [
@@ -33,7 +33,7 @@ setup(
             "pairinteraction-complex", "pairinteraction-real",
             "databases/quantum_defects.db"
         ],
-        "pairinteraction_gui" : [ "conf/example.pconf", "conf/example.sconf", "pairinteraction/icon.png" ],
+        "pairinteraction_gui" : [ "conf/example.pconf", "conf/example.sconf", "pairinteraction_app/icon.png" ],
     },
     scripts = [ "start_pairinteraction_gui" ],
     install_requires = install_requires,
